// apps/web/src/generated/image-manifest.ts
export const imageManifest = {
  "/assets/FOOD_PACKAGES/IMG_0101.jpg": {
    "src": "/_optimized/assets/FOOD_PACKAGES/IMG_0101.webp",
    "width": 3024,
    "height": 4032,
    "blurDataURL": "data:image/webp;base64,UklGRuoAAABXRUJQVlA4IN4AAACwBQCdASoSABgAPu1kq06ppaOiKA1RMB2JaACdMoMpA+gaxAcnElCV25+zytcnue1+Mzf+wAAA/tlhtac+eI7NBvlo964kMN1mmKpwhtFama5smGC+nlESyiN1GFUJcUiqtt4q+QbcFqPoONzaxX6o6vChFuS/DVpBdhHTpdxuCquSI8+RJJxKLPd9QIEZla1xZQijSMNujEn6IrvzB5CLntgBU431FncU10ZNkECmyCKn7ZOwcnRmK+RJAKcnLkPwqgaxbwWgBgfzDLhrHUlcLg4uj8VHxC0Y5livYAA="
  },
// ...
  "/assets/KATARAKT/JQ0G3807.jpg": {
    "src": "/_optimized/assets/KATARAKT/JQ0G3807.webp",
    "width": 5760,
    "height": 3840,
    "blurDataURL": "data:image/webp;base64,UklGRs4AAABXRUJQVlA4IMIAAADQBACdASoYABAAPu1iqU2ppaOiMAgBMB2JbACdMoEUwERWKo07Ge1nd2WjHQ2OAADxM2yKUMaxAIb694FJv56QNmrkE7/rftSsw21AtKtTb0C8IpEyOYyKTHMFoHb8MmL3ovEup2q+tBJRxgBflKf1CsaCeQyTPo66J4A4D2YXL8eUV/9r/1QniUg9AGLavfuBehHJ0Jw1c31dThAtxKQQxfs6BaS8EpaV+2V3ugGfUxmlsAWO6gmwW1Q/BhkKoAAAAA=="
  },
// ...
  "/assets/QUR_DISTRIBUTION/IMG_4337.JPG": {
    "src": "/_optimized/assets/QUR_DISTRIBUTION/IMG_4337.webp",
    "width": 3573,
    "height": 4466,
    "blurDataURL": "data:image/webp;base64,UklGRg4BAABXRUJQVlA4IAIBAAAwBgCdASoTABgAPu1qrlCppiQiqAqpMB2JbACdMoSA28AVxYaA6wr27DzWspVbpoUZbopzkzYcn4gAAP7GBARmCHuMfpU1/5IZy88wMvwLP3Xi7aKHN7BiYussdJSh5z6Y3f5ycSELbvfQyjablVmohSAL2jqFWXVFlgvjFolLrSjTVVUQpoMWqueq+hcxSjePBzqm2ahaEbR/6QnipkJejpsQUFsrG40MwdfxB394X0fthL+jSYxHsYxX31dZexfj50j9OzfZuIBKU+f4ww/UfQH74t9I/81BPAsXzQmJLFIj+AlBGLs7wqy+Q6qgD/3um8gZHXIsQhYjc4040a+AAAA="
  },
// ...
  "/hands.jpg": {
    "src": "/_optimized/hands.webp",
    "width": 1200,
    "height": 800,
    "blurDataURL": "data:image/webp;base64,UklGRtAAAABXRUJQVlA4IMQAAADwBACdASoYABAAPu1iqU2ppaOiMAgBMB2JZACdMoMYA+gayI5YYB1Fawxb6grBpAAA/od/0K+/vBFQUTbC3V9Dvj2EDevtw0qo/iWMaSJnjddvF+gj9zjxwra2cDPVWQHFf98IMAOYxb8tv05OHmsWewwDYNe2mAWpn+a++4E7PmTafEAhoq/6TH8Y5By39x2a0cj1Z8DmaxsiadCXHdax4iN6Vm0RcGm9+4JmcBExVvFWhAzKFHKbPXKaiUBFRw8RL3AA"
  },
// ...
  "/images/21.jpg": {
    "src": "/_optimized/images/21.webp",
    "width": 1350,
    "height": 1080,
    "blurDataURL": "data:image/webp;base64,UklGRhQBAABXRUJQVlA4IAgBAAAQBgCdASoYABMAPu1mqk8ppaOiKA1RMB2JbACdMt1w0Yu+gAAvhnfg7f3rFAMeBG7ayYZfT+DnW7AA/uf6lB9BDah8wME/VKdFUdx/CYCgMTYjIiC6Z1Y8AIf7IqpuMu2FsUsvE5ZqAL5NLu2tP5o2aRQnPb9WMR8/UMYN6eljGaNrwLTM9tcnvxKKxV0wDX/lAYKT5cJZTGpxWkKeiDPO4VU4teA+EMYJ3K6EAFaI7pVT4EuA6QBDKVfNuZMRsE3PugxtcX3puNnFkdrFGTXr2/3NSL9agcC5kHguQ+jV7g3awyDpn7VSg+sE97JpO5/teXw2oWmOzn2H8bi4CHAQSljA9UBQAAA="
  },
// ...
  "/images/IMG_2479_2.jpg": {
    "src": "/_optimized/images/IMG_2479_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRgYBAABXRUJQVlA4IPoAAADQBQCdASoTABgAPu1qqlEppaOiqAqpMB2JbAC1GzL6A0os74Oj6/H3T8a+vSMqRBkuAaHsoSzAAP6RuK0Ji7HazU7zPUaae45O7j+/zl/v2jaw+4rOnm48MPanqATcCGF/dJSyCBeeZZOaw3kTf0fCahi6gehkXJx1FZrtXRMUqM/3XnG12Qg0/k6oReQQo1sj9hdlX73WfV+0IcRzx2LMAA8snfToN2Au8KaWwaxay0VgsAnNLW3igSTh7VltndzPJyFHydsaB0673MzSetmdPcUZlt0BbJdpnOyaaiUtK3dD3TvmuwzbS6LBByEtjz+1VEvVdw5EwygA"
  },
// ...
  "/images/IMG_2512_2.jpg": {
    "src": "/_optimized/images/IMG_2512_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRhwBAABXRUJQVlA4IBABAABwBgCdASoTABgAPu1uqlEppiOiqAqpMB2JbACxHuN+HaA2wAp+AheG8Ok9DiUT7a/kaQdPKR4fnWEBcOAAzjWz4QALVcW/hEuHhpynZ7gFTrlUHpw8e5GJ2Ad7KqM99+SWZaiP4ZT0kbnW/ZnCXhMtuoQSD2FgMiT15Km3ROORW3BN1yKsEo2HQg0Ms/5csA9APwckWU4vA7ingr7hSSj9q/BMVfNNgS2gDysZ65lAgLdRHmIYfV/RjZXwuOHGjMrV1E1Kmc5uKw+lKLQ7+VTP1P+FGz6kEStI5qNgBqjNRItoYJBGTANZympbTZV9yEUArbzvDiskqvCHd77Y1ejYi6TB/aKOEvCMLnmNjiEAAA=="
  },
  "/images/IMG_2525_2.jpg": {
    "src": "/_optimized/images/IMG_2525_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRu4AAABXRUJQVlA4IOIAAADQBQCdASoTABgAPu1wsFIppiSiqAgBMB2JaACsMoO4AHTPGUQ3wq7bUaRG/bTwxBdGXilj1QgAAPaiylabzgib0pSNykKfC3WQffXS+cK45JA2XI8LtQAY46x415s/+kgAOLQrKzHgIKsyOKfN6OSRV+QVQWvyX1Reop9IAWNV9Dn1OzWDMC9oLHfJ4z5wjDAxRCcvvTa2KekMrVFBrvgMJoIGvPfIkQUvkEewgpv8mcrlhr5R5eZf8VMWFJ7eeuX5e8q18JQT9qA1rIVg3Eyc7j8kWxtFt8isazJqNETm7gAA"
  },
  "/images/IMG_2530_2.jpg": {
    "src": "/_optimized/images/IMG_2530_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRg4BAABXRUJQVlA4IAIBAABQBQCdASoTABgAPu1srVIppaQiqAgBMB2JbACuKA/yLDBmdEDZUQQsEp3i6cJ5jsZKxAAA/TsiDFTYhczQprWxU/lkzg3uhCSKm0d2WLapylCSu74vzCUSYLU76hw04rp0EOHYucEnGfOvMUYJv5Nsy8SEuDgPz96/VDBHAWqobcVtIrOshByYvnTWinB7r9mHCMgMTYC5MwCZ90BW+1q4vnx0TMKXM8F3UFYmNdQJIG5/AX0DqPid9883+roAAujIv2UdH4x4GH9AlKxlGfKQ2H4tGE38XHdeYyb3ApQRXut9litsWnuS7+Br1VaO95UMBlZYGdB6F0pcTEapEjcGAAA="
  },
// ...
  "/images/IMG_8983_2.jpg": {
    "src": "/_optimized/images/IMG_8983_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRg4BAABXRUJQVlA4IAIBAAAwBgCdASoTABgAPu1oqVAppaOiqA1RMB2JbACdMzRA2kgT4BsByFNvNfqfp5d2A8Iuge+f3FL7gJAAAP6veUgBY3yPvONY/XcGHWaFoLOLzJAMw6OtMqXw5xbHuk76xL1UAPPG3H1xJOYGk6hd+1TEnreYWn/S0mP+/snB7gJoK8YBzvYg9WRZWZSeTkhJH/eOl6bpYiFdiyRcn2QnHlLFzITqk7wgcjl9BYZTRe2Y8CeEEsN32OVzAMOl3p71f+Rdn8DcoZM8cgnBdZtvChhRVsedaeH4nRFyS3GzCv3dzyxpBX0B18FrFfqfBiXFzCsAGrgdxlhmtOmGoj0r/dQAAAA="
  },
  "/images/IMG_9003_2.jpg": {
    "src": "/_optimized/images/IMG_9003_2.webp",
    "width": 1600,
    "height": 2000,
    "blurDataURL": "data:image/webp;base64,UklGRtQAAABXRUJQVlA4IMgAAABwBQCdASoTABgAPu1sqlEppaOiqAqpMB2JZACxG46A5cM6nhFDLGW372TktqzjMEEwgkAAAP7U+UHgoYN7IsgtnX21tZvugqg7Fn0sfbg4F5+OGMv4Gxm6x45sJxdQ4nIQyYlbElOQfHsJKeURTa6BjniaJiRvN/nCDr2s7UYV+5gKpjvIC5YbfnTnKg+z3V27y4eF/DUMU3Rx7POtXv8a8zF2FiociQXM8nESbP79rB+930CKg9IDJIAV4nD2DjLcvbxBdHsAAA=="
  },
// ...
  "/images/IMG_9014_2.jpg": {
    "src": "/_optimized/images/IMG_9014_2.webp",
    "width": 2000,
    "height": 1500,
    "blurDataURL": "data:image/webp;base64,UklGRr4AAABXRUJQVlA4ILIAAAAwBACdASoYABIAPu1qq1EppaOiqAqpMB2JQBOkERgiL8HLx4lqc/HIwAAA4nN7CtHwXpsWoAzakmj6ydznqG3+KCueo96cdrcLO1Hsn8QFQD4Ev7FDzMArX3JTNQvzQ6+fy/t3EPj0fa/xeJr0wWzOfYlZJgN7WLZ7hSn9qbQG5+hmkL2LoH2QG7vDcrHT2guFycrbTVjDbh5/RMhqAUDmc9H6r1976tLHXT9NOCkkHgAA"
  },
// ...
  "/images/placeholder.jpg": {
    "src": "/images/placeholder.jpg",
    "width": 1200,
    "height": 800,
    "blurDataURL": "data:image/webp;base64,UklGRjYAAABXRUJQVlA4ICoAAADwAgCdASoYABAAPu1iqU2ppaOiMAgBMB2JZwDCgC0kAAD+76wyxuUAAAA="
  },
// ...
  "/placeholder.jpg": {
    "src": "/placeholder.jpg",
    "width": 1200,
    "height": 800,
    "blurDataURL": "data:image/webp;base64,UklGRjYAAABXRUJQVlA4ICoAAADwAgCdASoYABAAPu1iqU2ppaOiMAgBMB2JZwDCgC0kAAD+76wyxuUAAAA="
  }
} as const;